+++ previous-versions/accounting/src/codataccounting/models/operations/get_invoice_attachment.py
@@ -12,11 +12,13 @@ from typing import Optional
 @dataclasses.dataclass
 class GetInvoiceAttachmentRequest:
     attachment_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'attachmentId', 'style': 'simple', 'explode': False }})
-    r"""Unique identifier for an attachment"""
+    r"""Unique identifier for an attachment."""
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     connection_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'connectionId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a connection."""
     invoice_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'invoiceId', 'style': 'simple', 'explode': False }})
-    r"""Unique identifier for an invoice"""
+    r"""Unique identifier for an invoice."""
     
 
 
@@ -25,11 +27,14 @@ class GetInvoiceAttachmentRequest:
 @dataclasses.dataclass
 class GetInvoiceAttachmentResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     attachment: Optional[shared_attachment.Attachment] = dataclasses.field(default=None)
     r"""Success"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""