+++ previous-versions/accounting/src/codataccounting/models/shared/salesorderref.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 import dataclasses
+from ..shared import datatype as shared_datatype
 from codataccounting import utils
 from dataclasses_json import Undefined, dataclass_json
 from typing import Optional
@@ -11,7 +12,9 @@ from typing import Optional
 
 @dataclasses.dataclass
 class SalesOrderRef:
-    data_type: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('dataType'), 'exclude': lambda f: f is None }})
+    data_type: Optional[shared_datatype.DataType] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('dataType'), 'exclude': lambda f: f is None }})
+    r"""Available Data types"""
     id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('id'), 'exclude': lambda f: f is None }})
+    r"""Unique identifier to a record in `dataType`."""