+++ previous-versions/accounting/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [python v0.30.0] previous-versions/accounting
 ### Releases
 - [PyPI v0.30.0] https://pypi.org/project/codat-accounting/0.30.0 - previous-versions/accounting
+
+## 2023-10-16 01:24:52
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.99.2 (2.155.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.31.0] previous-versions/accounting
+### Releases
+- [PyPI v0.31.0] https://pypi.org/project/codat-accounting/0.31.0 - previous-versions/accounting